+++ app/controllers/Chofer.php
@@ -14,24 +14,29 @@ class Chofer extends Control
 
     public function index($errores = [])
     {
-        $choferes = $this->modelo->getAllChoferes();
-        $datos = [
-            'title' => 'Listado de Choferes',
-            'urlCrear' => URL . '/chofer/create',
-            'columnas' => ['Nombre', 'Apellido', 'DNI', 'Nacionalidad'],
-            'columnas_claves' => ['nombre', 'apellido', 'dni', 'nacionalidad'],
-            'data' => $choferes,
-            'acciones' => function($fila) {
-                $id = $fila['id_chofer'];
-                $url = URL . '/chofer';
-                return '
-                    <a href="'.$url.'/edit/'.$id.'" class="btn btn-sm btn-outline-primary">Editar</a>
-                    <a href="'.$url.'/delete/'.$id.'" class="btn btn-sm btn-outline-danger" onclick="return confirm(\'¿Eliminar este chofer?\');">Eliminar</a>
-                ';
-            },
-            'errores' => $errores
-        ];    
-        $this->load_view('partials/tablaAbm', $datos);
+        if (in_array('ver abm',$_SESSION['usuario_derechos'])) {
+            $choferes = $this->modelo->getAllChoferes();
+            $datos = [
+                'title' => 'Listado de Choferes',
+                'urlCrear' => URL . '/chofer/create',
+                'columnas' => ['Nombre', 'Apellido', 'DNI', 'Nacionalidad'],
+                'columnas_claves' => ['nombre', 'apellido', 'dni', 'nacionalidad'],
+                'data' => $choferes,
+                'acciones' => function($fila) {
+                    $id = $fila['id_chofer'];
+                    $url = URL . '/chofer';
+                    return '
+                        <a href="'.$url.'/edit/'.$id.'" class="btn btn-sm btn-outline-primary">Editar</a>
+                        <a href="'.$url.'/delete/'.$id.'" class="btn btn-sm btn-outline-danger" onclick="return confirm(\'¿Eliminar este chofer?\');">Eliminar</a>
+                    ';
+                },
+                'errores' => $errores
+            ];    
+            $this->load_view('partials/tablaAbm', $datos);
+        } else {
+            header("Location: " . URL);
+            exit;
+        }
     }
     
     public function edit($id)
