+++ app/controllers/ReservasPuntos.php
@@ -15,23 +15,28 @@ class ReservasPuntos extends Control
     // Mostrar todas las reservas
     public function index()
     {
-        $reservas = $this->model->getAllReservasPuntos();
-        $datos = [
-            'title' => 'Listado de Puntos Reservados',
-            'urlCrear' => URL . '/reservaspuntos/create',
-            'columnas' => ['Fecha', 'Hotel','Punto de Detencion'.'Permiso'],
-            'columnas_claves' => ['fecha_horario','hotel','punto_detencion','id_permiso'],
-            'data' => $reservas,
-            'acciones' => function($fila) {
-                $id = $fila['id_reserva_punto'];
-                $url = URL . '/reservaspuntos';
-                return '
-                    <a href="'.$url.'/edit/'.$id.'" class="btn btn-sm btn-outline-primary">Editar</a>
-                    <a href="'.$url.'/delete/'.$id.'" class="btn btn-sm btn-outline-danger" onclick="return confirm(\'¿Eliminar esta reservacion?\');">Eliminar</a>
-                ';
-            }
-        ];    
-        $this->load_view('partials/tablaAbm', $datos);
+        if(in_array('ver abm', $_SESSION['usuario_derechos'])) {
+            $reservas = $this->model->getAllReservasPuntos();
+            $datos = [
+                'title' => 'Listado de Puntos Reservados',
+                'urlCrear' => URL . '/reservaspuntos/create',
+                'columnas' => ['Fecha', 'Hotel','Punto de Detencion'.'Permiso'],
+                'columnas_claves' => ['fecha_horario','hotel','punto_detencion','id_permiso'],
+                'data' => $reservas,
+                'acciones' => function($fila) {
+                    $id = $fila['id_reserva_punto'];
+                    $url = URL . '/reservaspuntos';
+                    return '
+                        <a href="'.$url.'/edit/'.$id.'" class="btn btn-sm btn-outline-primary">Editar</a>
+                        <a href="'.$url.'/delete/'.$id.'" class="btn btn-sm btn-outline-danger" onclick="return confirm(\'¿Eliminar esta reservacion?\');">Eliminar</a>
+                    ';
+                }
+            ];    
+            $this->load_view('partials/tablaAbm', $datos);
+        } else {
+            header("Location: " . URL);
+            exit;
+        }
     }
 
     // Mostrar una reserva específica
